feat(app): rename hooks
