+++ examples/binds.ts
@@ -1,4 +1,4 @@
-import { WelcomeComponent } from './welcome'
+import { WelcomeComponent } from './pages/welcome/welcome.component'
 import { Logger } from './services/log.service'
 import { User } from './services/user.service'
 
@@ -6,4 +6,8 @@ export const binds = {
   app: WelcomeComponent,
   logger: Logger,
   user: User,
+}
+
+export class Module {
+
 }
