+++ examples/pages/welcome/welcome.component.html
@@ -1,0 +1,1 @@
+<p>hello</p>
