+++ examples/pages/welcome/welcome.component.ts
@@ -1,27 +1,27 @@
-import { Component } from '../src/annotations/component'
-import { Logger } from './services/log.service'
+import { Component } from '../../../src/annotations/component'
+import { Logger } from '../../services/log.service'
+// import template from './welcome.component.html'
 
 @Component({
-  props: {
-    propMessage: String,
-  },
-  components: {},
+  template: `
+  <p>...</p>
+  `,
 })
 export class WelcomeComponent {
-
+  
   propMessage: string
-
+  
   constructor(
     private logger: Logger,
   ) {
   }
-
+  
   mounted(): void {
     console.log('inject logger:', this.logger)
     this.greet()
   }
-
+  
   greet(): void {
   }
-
+  
 }
